+++ doc/whatsnew.rst
@@ -28,6 +28,7 @@ Breaking changes
 New features
 ~~~~~~~~~~~~
 
+- Add :meth:`xarray.Dataset.hyoga.assign`.
 - Add :meth:`xarray.Dataset.hyoga.assign_icemask`.
 - Add :meth:`xarray.Dataset.hyoga.where_icemask`.
 - Plot methods look for variable `land_ice_area_fraction`.